# openapi.yaml
openapi: 3.0.3
info:
  title: Kaldar Platform API Documentation
  description: |
    Complete Production REST API specification for the Kaldar Monorepo (Laundry & Dry Cleaning Platform).
    
    ### Full API Inventory (55 Endpoints across 14 Controllers):
    1. **Authentication & Verification**: Login, OTP verification & resending, Forgot Password, Reset Password.
    2. **Customer**: Registration, profile fetch, update profile, change password, delete customer account.
    3. **Dry Cleaner Store & Ops**: Registration, profile edit, CAC business verification & status check, service inventory management, payout bank account setup, working hours schedule, onboarding checklist status, analytics metrics, delete dry cleaner account.
    4. **Payout**: Bank account verification, earnings withdrawal to registered bank (Anchor Banking).
    5. **Discovery**: Geospatial discovery of nearby available dry cleaners.
    6. **Order Management**: Create order, accept order, reject order, order status progression, fetch order details, customer order history, dry cleaner order queue.
    7. **Review & Ratings**: Submit review, dry cleaner store reviews list, single order review fetch.
    8. **Logistics & Dispatch**: Delivery partner webhook callbacks (Chowdeck, Uber Direct, etc.).
    9. **Notification Center**: Paginated notifications list, toggle read status, delete single, bulk mark read/unread/delete, mark all read.
...
  version: 1.0.0
  contact:
    name: Kaldar Engineering Team
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/drycleaner/drycleaner/register:
    post:
      tags: [DryCleaner]
      summary: Register Dry Cleaner Store
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [businessName, email, password, phoneNumber, address]
              properties:
                businessName: { type: string }
                email: { type: string }
                password: { type: string }
                phoneNumber: { type: string }
                address: { type: string }
      responses:
        '201': { description: Dry cleaner registered }
  /api/v1/drycleaner/drycleaner/edit-profile:
    patch:
      tags: [DryCleaner]
      summary: Edit Dry Cleaner Store Profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Profile edited }
  /api/v1/drycleaner/drycleaner/verify-business:
    post:
      tags: [DryCleaner]
      summary: Submit CAC Business Verification
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data: { type: object }
                cacDocument: { type: string, format: binary }
      responses:
        '201': { description: Verification submitted }
  /api/v1/drycleaner/drycleaner/{dryCleanerId}/verification-status:
    get:
      tags: [DryCleaner]
      summary: Get CAC Verification Status
      parameters:
        - name: dryCleanerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Status retrieved }
  /api/v1/drycleaner/drycleaner/{dryCleanerId}/services:
    get:
      tags: [DryCleaner]
      summary: Get Dry Cleaner Services Offering List
      parameters:
        - name: dryCleanerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Services list }
    post:
      tags: [DryCleaner]
      summary: Add or Update Service Offering
      parameters:
        - name: dryCleanerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceName, price]
              properties:
                serviceName: { type: string }
                price: { type: number }
      responses:
        '201': { description: Service offering added/updated }
  /api/v1/drycleaner/drycleaner/payout-account:
    post:
      tags: [DryCleaner]
      summary: Update Bank Payout Account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dryCleanerId, accountName, accountNumber, bankCode, bankName]
              properties:
                dryCleanerId: { type: integer }
                accountName: { type: string }
                accountNumber: { type: string }
                bankCode: { type: string }
                bankName: { type: string }
      responses:
        '200': { description: Payout account saved }
  /api/v1/drycleaner/drycleaner/working-hours:
    post:
      tags: [DryCleaner]
      summary: Update Store Working Hours
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dryCleanerId, workingHoursJson]
              properties:
                dryCleanerId: { type: integer }
                workingHoursJson: { type: string }
      responses:
        '200': { description: Working hours saved }
  /api/v1/drycleaner/drycleaner/{dryCleanerId}/onboarding-status:
    get:
      tags: [DryCleaner]
      summary: Get Onboarding Checklist Progress
      parameters:
        - name: dryCleanerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Onboarding status fetched }
  /api/v1/drycleaner/drycleaner/{dryCleanerId}/analytics:
    get:
      tags: [DryCleaner]
      summary: Get Store Performance Analytics
      parameters:
        - name: dryCleanerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: period
          in: query
          schema: { type: string, default: "today" }
      responses:
        '200': { description: Analytics metrics fetched }
  /api/v1/drycleaner/drycleaner/{dryCleanerId}:
    delete:
      tags: [DryCleaner]
      summary: Delete Dry Cleaner Account
      parameters:
        - name: dryCleanerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { description: Dry cleaner account deleted successfully }
        '404': { description: Dry cleaner not found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT